edit song, edit lyrics works

+++ SetSelector/components/editlibrary.js
@@ -1,7 +1,7 @@
 
 const editLibrary_html = `
 <div id="popup-editlibrary" class="popup">
-    <p class="section_label" style="font-size: 1.3rem"> Personal Song Library</p>
+    <p class="section_label" style="font-size: 1.3rem"> Song Library</p>
     <img class="popup_close" src="icons/x-circle-fill.svg" onclick="closeEditSongLibraryPopUp()">
     <div id="popup-personal_search_content" class="search_bar_content">
     <ul id="popup-personal_search_results" class="search_results"></ul>
@@ -29,6 +29,28 @@ const editLibrary_html = `
     </div> 
 </div>
 `
+
+
+const editSong_html = `
+<div id="popup-editsong" class="popup">
+    <p id="popup-title" class="section_label"> Edit Song </p>
+    <img class="popup_close" src="icons/x-circle-fill.svg" onclick="closeEditSongPopUp()">
+        <label for="edit_song_title_input" class="form_label">Title</label>
+        <input type="text" id="edit_song_title_input" name="song_title_input" disabled="disabled" class="popup_input" required>
+        <label for="edit_song_author_input" class="form_label">Author</label>
+        <input type="text" id="edit_song_author_input" name="song_author_input" class="popup_input" required>
+        <label for="edit_song_tempo_input" class="form_label">Tempo</label>
+        <select id="edit_song_tempo_input" class="popup_input" required>
+            <option value="Fast">Fast</option>
+            <option value="Medium">Medium</option>
+            <option value="Slow">Slow</option>
+            <option value="Intercessory">Intercessory</option>
+        </select>
+        <label for="edit_song_sheet_input" class="form_label" required>Chord Sheet</label>
+        <textarea id="edit_song_sheet_input" class="popup_input" placeholder="Use two blank lines to separate verse/chorus/etc."></textarea>
+        <button id="song_input_submit" class="popup_input" onclick="editSongInLibrarySubmit()"> Submit </button>
+</div>
+`
 function editSongLibraryPopUp(){
     document.getElementById("editlibrarypopupplaceholder").style.display = "block";
     document.getElementById("popup-background").style.display = "block";  
@@ -39,31 +61,87 @@ function editSongLibraryPopUp(){
     document.getElementById("popup-background").style.display = "none";
   }
   
+  function editSongPopUp() {
+    closeEditSongLibraryPopUp();
+
+    document.getElementById("editsongpopupplaceholder").style.display = "block";
+    document.getElementById("popup-background").style.display = "block";
+
+    let target_title = this.parentNode.parentNode.firstChild.innerText;
+    let song_database = JSON.parse(localStorage.getItem("song_database"));
+  
+    for (let i=0; i<song_database.length;i++) {
+      if (song_database[i].title == target_title) {
+  
+          let song = song_database[i];         
+          document.getElementById("edit_song_title_input").value = song.title;
+          document.getElementById("edit_song_author_input").value = song.author;
+          document.getElementById("edit_song_tempo_input").value = song.tempo;
+          document.getElementById("edit_song_sheet_input").value = song.sheet;
+      }
+    }
+  
+  }
+
+  function closeEditSongPopUp() {
+    document.getElementById("editsongpopupplaceholder").style.display = "none";
+    document.getElementById("popup-background").style.display = "none";
+
+    editSongLibraryPopUp();
+
+  }
+
   function closeAllPopUp() {
     for (let node of document.getElementsByClassName('popup-placeholder')) 
       node.style.display = "none";
   }
 
-  function removePersonalSong() {
+  function removeSongInLibrary() {
+    if (confirm("Remove Song from Library?")) {
 
-    let target_title = this.parentNode.parentNode.firstChild.innerText;
-    console.log(target_title)
+      let target_title = this.parentNode.parentNode.firstChild.innerText;
+      console.log(target_title)
 
-    let song_database = JSON.parse(localStorage.getItem("song_database"));
+      let song_database = JSON.parse(localStorage.getItem("song_database"));
 
-    for (let i=0; i<song_database.length;i++) {
-      console.log(song_database[i].title);
+      for (let i=0; i<song_database.length;i++) {
+        console.log(song_database[i].title);
 
-      if (song_database[i].title == target_title) {
-          song_database.splice(i, 1);
-          localStorage.setItem("song_database", JSON.stringify(song_database));
-          console.log(target_title + " removed");
+        if (song_database[i].title == target_title) {
+            song_database.splice(i, 1);
+            localStorage.setItem("song_database", JSON.stringify(song_database));
+            console.log(target_title + " removed");
 
-          reloadDatabase();
+            reloadDatabase();
+        }
       }
-    }
+  }
 }
 
+
+function editSongInLibrarySubmit() {
+  let target_title = document.getElementById("edit_song_title_input").value;
+
+  let song_database = JSON.parse(localStorage.getItem("song_database"));
+
+  for (let i=0; i<song_database.length;i++) {
+
+    if (song_database[i].title == target_title) {
+
+        let song = song_database[i];
+        song.author = document.getElementById("edit_song_author_input").value;
+        song.tempo = document.getElementById("edit_song_tempo_input").value;
+        song.sheet = document.getElementById("edit_song_sheet_input").value;
+
+        localStorage.setItem("song_database", JSON.stringify(song_database));
+        reloadDatabase();
+    }
+  }
+
+  closeEditSongPopUp();
+
+
+}
 
 function resetLocalDatabase(){
   if (confirm("You will lose all of your edits/added songs and reset the library to the default one. Proceed?")) {
@@ -107,4 +185,11 @@ function reloadEditLibrary() {
     document.getElementById("editlibrarypopupplaceholder").innerHTML = editLibrary_html;
   }
   
+function reloadEditSong() {
+  document.getElementById("editsongpopupplaceholder").innerHTML = editSong_html;
+
+
+}
+
   reloadEditLibrary(); 
+  reloadEditSong();
